# backend/docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/operations/blockchain/reconcile": {
      "post": {
        "operationId": "BlockchainOperationsController_reconcile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockchainReconciliationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/licenses": {
      "get": {
        "operationId": "LicenseController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LicenseProjectionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licenses"
        ]
      }
    },
    "/api/v1/licenses/{id}": {
      "get": {
        "operationId": "LicenseController_find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseProjectionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licenses"
        ]
      }
    },
    "/api/v1/licenses/{id}/activation-key/retrieve": {
      "post": {
        "operationId": "LicenseController_retrieve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationKeyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a confirmed activation key exactly once",
        "tags": [
          "licenses"
        ]
      }
    },
    "/api/v1/public/licenses/{publicId}/verify": {
      "get": {
        "operationId": "PublicLicenseController_verify",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicLicenseVerificationDto"
                }
              }
            }
          }
        },
        "summary": "Verify an allowlisted public license projection",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/orders": {
      "post": {
        "operationId": "CommerceController_create",
        "parameters": [
          {
            "name": "X-License-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an authenticated idempotent purchase or renewal order",
        "tags": [
          "commerce"
        ]
      },
      "get": {
        "operationId": "CommerceController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "CommerceController_find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/orders/{id}/terms": {
      "get": {
        "operationId": "CommerceController_terms",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderTermsDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the exact Terms artefact snapshotted by an order",
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/orders/{id}/accept-terms": {
      "post": {
        "operationId": "CommerceController_acceptTerms",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptTermsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/orders/{id}/checkout": {
      "post": {
        "operationId": "CommerceController_checkout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "post": {
        "operationId": "CommerceController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/payments/ipn": {
      "post": {
        "operationId": "PaymentController_ingest",
        "parameters": [
          {
            "name": "x-emukey-payment-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIngestResultDto"
                }
              }
            }
          }
        },
        "summary": "Ingest an idempotent payment provider event",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/history": {
      "get": {
        "operationId": "PaymentController_history",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentHistoryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List role-scoped durable payment evidence",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/review": {
      "get": {
        "operationId": "PaymentController_reviewQueue",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentReviewDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/review/{id}": {
      "post": {
        "operationId": "PaymentController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentReviewDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/{id}/receipt": {
      "get": {
        "operationId": "PaymentController_getReceipt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentReceiptDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a receipt derived from durable matched payment evidence",
        "tags": [
          "payments"
        ]
      }
    }
  },
  "info": {
    "title": "EmuKey API",
    "description": "Versioned API for the EmuKey license platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BlockchainReconciliationDto": {
        "type": "object",
        "properties": {
          "health": {
            "$ref": "#/components/schemas/BlockchainHealthDto"
          },
          "indexedEvents": {
            "type": "number"
          },
          "processed": {
            "type": "boolean"
          },
          "projection": {
            "$ref": "#/components/schemas/BlockchainProjectionRepairDto"
          },
          "reconciledCommandIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "health",
          "indexedEvents",
          "processed",
          "projection",
          "reconciledCommandIds"
        ]
      },
...
      "LicenseProjectionDto": {
        "type": "object",
        "properties": {
          "blockNumber": {
            "type": "number",
            "nullable": true
          },
          "confirmationCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "entitlementVersion": {
            "type": "number"
          },
          "expiresAt": {
            "type": "string"
          },
          "finality": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "keyVersion": {
            "type": "number"
          },
          "maxActiveDevices": {
            "type": "number"
          },
          "originOrderId": {
            "type": "string",
            "format": "uuid"
          },
          "periodStart": {
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/LicensePlanDto"
          },
          "productName": {
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/LicenseProviderDto"
          },
          "publicLicenseId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_ONCHAIN",
              "ACTIVE",
              "SUSPENDED",
              "EXPIRED",
              "REVOKED"
            ]
          },
          "transactionHash": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "confirmationCount",
          "createdAt",
          "entitlementVersion",
          "expiresAt",
          "finality",
          "id",
          "keyVersion",
          "maxActiveDevices",
          "originOrderId",
          "periodStart",
          "plan",
          "productName",
          "provider",
          "publicLicenseId",
          "status",
          "updatedAt"
        ]
      },
      "ActivationKeyDto": {
        "type": "object",
        "properties": {
          "activationKey": {
            "type": "string"
          },
          "keyVersion": {
            "type": "number"
          }
        },
        "required": [
          "activationKey",
          "keyVersion"
        ]
      },
      "PublicLicenseVerificationDto": {
        "type": "object",
        "properties": {
          "blockNumber": {
            "type": "number",
            "nullable": true
          },
          "confirmationCount": {
            "type": "number"
          },
          "expiresAt": {
            "type": "string"
          },
          "finality": {
            "type": "string"
          },
          "licenseId": {
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/LicensePlanDto"
          },
          "productName": {
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/LicenseProviderDto"
          },
          "state": {
            "type": "string",
            "enum": [
              "CHAIN_CONFIRMED",
              "PENDING_ONCHAIN",
              "PROJECTION_STALE",
              "REORGED",
              "NOT_FOUND"
            ]
          },
          "status": {
            "type": "string"
          },
          "transactionHash": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "state"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "format": "uuid"
          },
          "targetLicenseId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "planId"
        ]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "billingCycleSnapshot": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "currency": {
            "type": "string"
          },
          "customerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "durationMonthsSnapshot": {
            "type": "number"
          },
          "entitlementsSnapshot": {
            "type": "object"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "licenseId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "maxActiveDevicesSnapshot": {
            "type": "number"
          },
          "orderNumber": {
            "type": "string"
          },
          "orderStatus": {
            "type": "string",
            "enum": [
              "WAITING_TERMS_ACCEPTANCE",
              "WAITING_PAYMENT",
              "PAYMENT_ACCEPTED",
              "CANCELLED",
              "EXPIRED"
            ]
          },
          "orderType": {
            "type": "string",
            "enum": [
              "NEW_PURCHASE",
              "RENEWAL"
            ]
          },
          "paymentDueAt": {
            "type": "string",
            "format": "date-time"
          },
          "planCommitmentSnapshot": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "planId": {
            "type": "string",
            "format": "uuid"
          },
          "planNameSnapshot": {
            "type": "string"
          },
          "planVersionSnapshot": {
            "type": "number"
          },
          "priceVndSnapshot": {
            "type": "number"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productNameSnapshot": {
            "type": "string"
          },
          "providerNameSnapshot": {
            "type": "string"
          },
          "providerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "publicLicenseId": {
            "type": "string",
            "nullable": true
          },
          "targetLicenseId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "termsAcceptedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "termsHashSnapshot": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "termsVersionSnapshot": {
            "type": "number"
          }
        },
        "required": [
          "billingCycleSnapshot",
          "createdAt",
          "currency",
          "customerUserId",
          "durationMonthsSnapshot",
          "entitlementsSnapshot",
          "id",
          "maxActiveDevicesSnapshot",
          "orderNumber",
          "orderStatus",
          "orderType",
          "paymentDueAt",
          "planCommitmentSnapshot",
          "planId",
          "planNameSnapshot",
          "planVersionSnapshot",
          "priceVndSnapshot",
          "productId",
          "productNameSnapshot",
          "providerNameSnapshot",
          "providerUserId",
          "termsHashSnapshot",
          "termsVersionSnapshot"
        ]
      },
      "OrderTermsDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "hash": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "content",
          "hash",
          "version"
        ]
      },
      "AcceptTermsDto": {
        "type": "object",
        "properties": {
          "termsVersion": {
            "type": "number",
            "minimum": 1
          },
          "termsHash": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          }
        },
        "required": [
          "termsVersion",
          "termsHash"
        ]
      },
      "CheckoutSessionDto": {
        "type": "object",
        "properties": {
          "amountVnd": {
            "type": "number"
          },
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "checkoutReference": {
            "type": "string"
          },
          "checkoutUrl": {
            "type": "string",
            "format": "uri"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresWithOrder": {
            "type": "boolean"
          }
        },
        "required": [
          "amountVnd",
          "attemptId",
          "checkoutReference",
          "checkoutUrl",
          "expiresAt",
          "expiresWithOrder"
        ]
      },
      "PaymentIngestResultDto": {
        "type": "object",
        "properties": {
          "activationRequired": {
            "type": "boolean"
          },
          "classification": {
            "type": "string",
            "enum": [
              "MATCHED",
              "DUPLICATE",
              "UNMATCHED",
              "AMOUNT_MISMATCH",
              "INVALID"
            ]
          },
          "commandId": {
            "type": "string",
            "format": "uuid"
          },
          "licenseId": {
            "type": "string",
            "format": "uuid"
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "classification",
          "transactionId"
        ]
      },
      "PaymentHistoryDto": {
        "type": "object",
        "properties": {
          "amountVnd": {
            "type": "number"
          },
          "classification": {
            "type": "string",
            "enum": [
              "MATCHED",
              "DUPLICATE",
              "UNMATCHED",
              "AMOUNT_MISMATCH",
              "INVALID"
            ]
          },
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string"
          },
          "orderType": {
            "type": "string",
            "enum": [
              "NEW_PURCHASE",
              "RENEWAL"
            ]
          },
          "planNameSnapshot": {
            "type": "string"
          },
          "productNameSnapshot": {
            "type": "string"
          },
          "providerEventId": {
            "type": "string"
          },
          "providerTransactionReference": {
            "type": "string",
            "nullable": true
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewStatus": {
            "type": "string",
            "nullable": true
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "amountVnd",
          "classification",
          "orderId",
          "orderNumber",
          "orderType",
          "planNameSnapshot",
          "productNameSnapshot",
          "providerEventId",
          "receivedAt",
          "transactionId"
        ]
      },
...
      "ReviewPaymentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "RESOLVED",
              "CLOSED_NO_ACTION"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "reason"
        ]
      },
...
    }
  }
}